Fix test-tests-1
TEST_WITH_MESSAGE prints the failing condition
Some functions have been renamed to indicate the state they are called in... maybe not great but more descriptive.
Now pop_pending_tests does not decrease count when count is already zero. This is a thorny issue: are we really going to go past the pop of zero?
In a way yes, in a way no: at the moment we don't because we test count before we pop, and we only pop when count is >0.
So, doing the extra test for no reason in the pop function is not good.
This is where we need to study assembler to see whether it is easier to just build a guard.
The current design of queue is that it is more like a ring with a placeholder element so that we don't need branching, on the assumption that is better than having a test every time we add or remove elements.